Disclaimer clarification
The disclaimer is not legally part of the license. I added it to clarify any confusion and make it clear that you are not allowed to re-license under my artist name "Kaelygon" as this would be impersonation and I don't want to be liable for any derivations that aren't created by me. 

You are of course free to remove the licenses completely and release as you see fit